Update s_advsound.cpp
Improved include file handling.

Specifies that SNDINFO includes with no leading "." are always relative to the containing directory of the including file (implicitly "./").

Handles explicit "./" and also ".." references.

Including across archives is not supported; it would be an odd thing to do with SNDINFO and can't think of a valid use case. Maybe in the future.